Updated the endpoint path for item/category
