+++ backend/src/routes/root-route.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const authRoute = require('./auth');
 const itemRoute = require('./item');
+const categoryRoute = require('./category');
 const verifyToken = require('../verify-token');
 
 const router = express.Router();
@@ -11,5 +12,6 @@ router.get('/', (req, res) => {
 
 router.use('/auth', authRoute);
 router.use('/item', [verifyToken, itemRoute]);
+router.use('/category', [verifyToken, categoryRoute])
 
 module.exports = router;
